schneider-openembedded: only run jobs on rzn1d
Rename the rzn1 template directory to rzn1d and make sure the jobs only
run on rzn1d, not on other boards such as soca9, which are not present
in LAVA yet.

And when soca9 comes online, only a subset of the tests will be
relevant. For example, soca9 doesn't boot a FIT image.

Change-Id: Ib22d77182dc48a68375bc462f170a86539a21cb4
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>