zephyr-upstream: Use ZEPHYR_BASE instead of WORKSPACE.
Don't assume any particular relation of where Zephyr is checked out vs
Jenkins-supplied WORKSPACE var. This is in preparation in migration to
west-imposed directory structure, which requires intermediate levels
between WORKSPACE and ZEPHYR_BASE. Overall, this change on its own
is intended to be a null change, to simplify review and testing of
further changes.

Change-Id: I7c648b6c36f421f62d18a396557d5d9edf3bd3e6