openembedded-lkft: submit_for_testing: Allow for QA build suffix
In the case of the developer build, a suffix needs to be added
in order to avoid a clash of builds reporting in the same
place. For example: Developer A builds mainline, and then
Developer B also builds the same revision of mainline; without
the distinctive suffix, both test results would end up in the
same QA build.

Change-Id: Iebf4358b5f6d8a34bd5758b058c2e9acb308d7a1
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>