openembedded-lkft: Add SDK job
This job will create a self-contained build environment that
can be used independently of OpenEmbedded, which will allow
for seamless cross-compilation.

The SDK's created are for:
* arm (thanks to x15 build)
* arm64 (via Juno build)
* x86 (from intel-corei7-64)

These SDK's can then be used to create Docker images for
cross-compilation using exactly the same toolchain that our OE
is using (such as GCC 7.3.0, Glibc 2.27, to name a few).

Change-Id: I3754c4f0a27f0cc96e1a2833fb02cc08bc73be4e
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>