lt-qcom-bootloader/dragonboard820c: generate ptable using Qualcomm ptool
Instead of using prebuilt versions of ptable, we now generate them
using partition.xml file as input. ptool.py is Qualcomm tool that is
used in production, and a new version was recently published (open
source) which allows anyone to generate such files. It makes
customization much simpler, since partition.xml can be edited and the
files can be regenerated.

Note that for UFS disks, it is not recommended to change the LUNs
layout, only ptable updates within each LUN should be attempted.

Change-Id: Ieb0930e2929f0b81e388a23ead50cbb739a9aab6
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>