ART: Job to build and package ART dependencies
Builds and packages dependencies to run ART tests. (and possibly
benchmarks too).
The package is uploaded to the Linaro release server.
This is set to run once nightly for the time being.

This job will in time replace part of the target tests job.
It uses the master-art branch of the manifest instead of master, which
syncs slightly more than 100 projects instead of full Android.

This job doesn't use the usual setup scripts in an attempt to simplify
the process for building and testing.

Testing will be done on other executors, freeing execution time for
builds on the build machines.

Change-Id: Ia4d6cdcf38e9114a8bf815097e0160348259b619