openembedded-lkft: Do not post build status if fetch failed
If there is a do_fetch failure due to a Git revision not
existing on the given branch (happens very often in Greg's
stable-rc branches), then there is no point in publishing the
build failure in QA Reports, as that would trigger an email
which wastes Greg's time in finding out why the build failed.

Greg's comment about this was:
  When your infra fails to fetch the source, please don't
  report this as a failure to us, as it wastes my time in
  digging through the web site to try to figure out what I
  broke in the build, when I did no such thing :(

This stops the script (and the build) if a line containing
  Fetcher failure: Unable to find revision
is found in the build logs.

Change-Id: Ifbe21814a952c0de27961126c8e3e8b7bd1ebea8
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>