lt-qcom-bootloader: include empty/zero boot image in the rescue package
When we reset the board with the rescue package, we also need to erase
the boot partition to make sure that there is no left out boot image
from the past. So let's create a small zero file that we will write to
boot partition, 1MB will be enough to clear the boot image header and
invalidate the boot data.

Change-Id: I50dd7c120b1d74cda407fc6394ea2084480b5185
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>