tcwg-benchmark: Change names of target_list values
Change from just 'tx1 | tk1' to 'tx1_64 | tx1_32 | tk1_32', so we can
differentiate between AArch64 and AArch32 runs.

Also update all jobs that may launch this one to pass in the correct
parameters. tcwg-llvm-patch-benchmarking is kind of stupid since it
expects targets that look a lot like labels and forwards apm_64 to tx1.
We should change that in a separate patch.

Change-Id: I9a9ba9e3bba84c668c1b21069cc16a78a6af8d11