openembedded-lkft: Allow for SSTATE_BRANCH override
When SSTATE_BRANCH is set, that directory will be used for
the sstate-cache. The purpose of this is to allow for other
manifests to be tested but referring to one of the big caches.

For example, if the oe-rpb-manifest's branch is 'd/my-test',
the link to the sstate-cache will include that branch name,
which will be empty in pretty much all cases. If, instead,
the SSTATE_BRANCH is set to 'lkft/sumo' (upon which the
oe-rpb-manifest is based, one can infer), then the build
will benefit from the already populated cache.

This speeds up distro-wide testing of changes.

Change-Id: I7dce6e5c47327943529d0b75ee195379d78bdaf2
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>