lt-qcom-debian-images: cleanup kernel command line parameters
-Remove rootwait, when using an initramdisk, the kernel doesn't
parse this. The debian initramdisk also doesn't parse this.
(But it still always waits for async devices to show up.)

-Remove rw. As long as /etc/fstab exists, systemd-fstab-generator
will create -.mount, and systemd-remount-fs.service will
remount all filesystems according to fstab, so the
ro/rw set by the initramdisk doesn't matter.

-Use PARTLABEL rather than /dev/disk/by-partlabel/.
PARTLABEL is supported by the debian initramdisk,
and it is the same syntax that is used by the kernel.
PARTLABEL is also more compact.

Change-Id: If19123cc62c43c0dcc9f8bbed07cb1c917b7872c
Signed-off-by: Niklas Cassel <niklas.cassel@linaro.org>