schneider-openembedded: cope with eth0/eth1 swap
Different kernels use different ethernet devices as their main port:

- 4.9  uses eth0
- 4.19 uses eth1

To cope with this, attempt to get the IP address from eth1, if it fails,
try eth0.

Also, add option "--touch" to not extract the timestamp from the test
overlay tarball, because the clocks on the DUT may not match the LAVA
server, giving a warning about the timestamp being in the future.

Change-Id: I8fe941726b8b619eff2daa0cb0aff7b9636be9ac
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>