zephyr-upstream: Upgrade gnuarmemb toolchain, and d/l it to persistent loc
Toolchains are location at ${HOME}/srv/toolchain/, which is backed by
a persistent Docker volume. Originally, new versions of toolchains
were installed to this volume manually on the host, but we can as
well do it "online" during the build, as long as we take care to
avoid downloading one if it's already available (i.e. was downloaded
previously and cached).

Based on the suggestion by Fathi Boudra.

Change-Id: I7543682cce9b6ace68134c2a421845b8426f2997
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>