LKFT: Implement lxc caching
Reuse lxc containers on dispatchers. This is slightly risky because it
will be cached under the lava-lxc name (so if the template changes but
the name doesn't, jobs could use a stale container). There are no great
ways to automatically decide when to invalidate the cache, so instead it
is hopefully well commented.

Tested @ https://lkft.validation.linaro.org/scheduler/job/1038403

Change-Id: Id04bfe60f9543e3ff0c2c405487e1209cc8d58c2
Signed-off-by: Dan Rue <dan.rue@linaro.org>