lt-qcom-linux-*: set ROOTFS and RAMDISK url in the actual job not the trigger
Currently, the trigger jobs are checking if there is a new kernel and
are also responsible to find the ramdisk and rootfs to use for the
testing. This can be done in the actual job instead of the
trigger. The test job should be able to control which ramdisk or
rootfs it needs.

As such the following changes were needed:
- remove RAMDISK_URL and ROOTFS_URL from the test jobs paramaters, and
from the triggers
- extract the function from get_latest_kernel_build that locates the
ramdisk/rootfs into its own script

Change-Id: Ic1e74a0cdd3cace488850256699a75df1cb63cc6
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>