se: fix docker test in UBI edge
The dip-image-edge UBI test fails to run docker because /tmp/sdcard is
not mounted as an ext4 image by default.

This is because we aren't flashing our own disk image for this test, but
relying on whatever is on the card from the last time the board was
used.

Instead of creating an ext4 partition on mmcblk0p1, just use tmpfs. For
real use-cases, this won't be good enough, but for a simple hello-world
test, there will be enough space.

Change-Id: Ifb4f70eee92080cbd265d16902a6567a53895cba
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>