lite-aeolus-openamp: Pass the actual published artifact to LAVA.
Previously, a sample image hosted on people.linaro.org was passed to LAVA.
We still use that image, but we use the URL of that image published as a
build artifact, so that gets us closer to the "real" CI job. Literally,
once that is working, the image build phase is the only missing part.

Change-Id: Ibeef69ea9290e1e4c5b7d0c6c6c2243a314fb6f4
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>