lite-iot-ci.yaml: Standalone definition for "lite-iot-ci" Jenkins view
Currently, a view is usually defined in a job definition file. But a
view applies to multiple jobs, which effectively means that either view
definition needs to be duplicated in every job which belongs to that view,
or a situation, that adding a new job, you need to patch completely
different job (which contains the view definition). This is currently
mostly alleviated using regex-based views. But the "lite-iot-ci" view
dealt with here is original an enumerated view. While it might be
possible, to re-cast it as an regex view, such a regex may grow unwieldy
anytime.

So, this is an attempt to structure job vs view definitions a bit better,
by introducing a standalone view definition files, like this.

Change-Id: Ibb52a66d7d12e323e6b1ed3c3739594dfb82c512
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>