se: YAML consolodation
Re-org the job submission YAMLs for Schneider.

Use the new facility in LAVA to allow multiple deployments and boots in
one job. Group the test YAMLs with common image variants, eg. dip-image,
dip-image-dev and dip-image-edge, as well as a specific YAML for the
LTP/PTEST job that runs for a long time.

Change-Id: I944b9660d42872525cfcd8286412c94c74f28ffc
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>