tcwg_bmk-*: Split into tcwg_bmk_tk1-* and tcwg_bmk_tx1-*
... to allow scaling the number of benchmarking board types.

Right now we are blocking all tcwg_bmk-* jobs when bisecting, which
causes TX1 (aarch64) jobs to block on TK1 (armhf) regressions and
vice versa.  This is bearable now, but with adding SynQuacers
(for SPEC CPU2017 benchmarkins) and/or X15 (for armhf) benchmarking
we will hit the bottleneck.

This also allows benchmarking of AArch32 code on ARMv8 boards.

Change-Id: I4fc9444f8c62d11baee14f364740ca36c7600f3d