lhg-oe-build: am57xx-evm: use template-igt-test regardless of MANIFEST_BRANCH
Fixes jenkins job failure for am57xx-evm thud builds.

Currently for am57xx-evm the morty builds are not tested in LAVA, the rocko
builds use template-igt-test.yaml, and the thud builds try to use the
default templates list (defined in lhg-oe-build/submit_for_testing.sh):
  template-robotframework-tests.yaml template-wifi-tests.yaml,
but these templates don't exist for this target.
Remove the 'if [ "${MANIFEST_BRANCH}" == "rocko" ]' check for am57xx-evm,
so that template-igt-test.yaml would always be used for this target.

Change-Id: I317f9b533ff5824a6e3647ecf8816ec38d3830d8
Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>