openembedded-lkft: Support filtering out tests migrated to LKFT 2.0
The following tests are already part of the lkft-full test
plan from lava-test-plans:
* kvm-unit-tests
* libhugetlbfs
* LTP
* spectre-meltdown-checker
* v4l2-compliance
so we need to remove them from here for the combinations that
have been migrated to LKFT 2.0.

The list of migrated combinations is in the form of
  QA_PROJECT::LAVA_DEVICE
one per line, which we can check for during test submission.

Change-Id: I9823ea68c58c2ca89eb0b2e5a8001480e3955bb3
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>