zephyr-upstream: Use flock to avoid race conditions installing SDK version
As this job runs as a matrix job, in case Zephyr SDK needs upgrade, there's
a case that multiple processes may write to a single, global SDK location
path, leading to weird race conditions. Normally, this was dealt with
uncompressing tarball to a temporary path, and then atomically mv'ing to
the final location, but in the case of Zephyr SDK, turned out that installer
patches installed ELF files to refer to the intermediate location, causing
them not to work when moved to the final location. So, instead switch to
use explicit locking using "flock" utility.

Change-Id: I7f53733bbf9de6ec64fc44c1c3ec0ff77faf6aa7
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>