tcwg_gnu.yaml: Mark build unstable if we skip a known bad config
This way we can tell at a glance what actually
needs fixing. It's not ideal because unstable
usually means test failures. However these jobs
seem to only completely pass or fail.

Also indent the same sections in bmk and
kernel files so that they can still be compared
for refactoring.

Change-Id: Icc7ff0ef504021119ccc318ef6bd4c6e0f9fe39b