lite-aeolus: Remove superfluous "job_name" field from LAVA metadata.
The initial reason for looking into this was inconsistency in naming:
"job_name", undescore as a seprator while the rest of metadata fields
use hyphen ("build-url", etc.)

But I decided to do more homework on the claim in comment ("For some
reason, LAVA doesn't allow to query by real job name"), and turned
out, that's actually not true. The whole confusion arised because
"for some reason", what's called "job_name" in the job definition
YAML, is called "description" in the LAVA UI.

The terminologi discrepancy is being reported to LAVA maintainer, but
in the meantime, the best action is to remove the superfluous field
which was added effectively by misconception.

Change-Id: Ib7f6217c0268f2b5b54830673baed972c840adbd
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>