trigger-linux-stable-rc: Git-describe using 12-hexdigits
Somewhat recently, the output of git-describe went from using
12-chars to 13-chars for the SHA. So, instead of
  v5.6.7-168-g853ae83af7cc
Git started to descrbe that same commit as
  v5.6.7-168-g853ae83af7cc6

That's not a problem for Git, there is no ambiguity, but we
use the git-description as the build ID in QA Reports, so for
the same git-commit we now have 2 different builds:
  https://qa-reports.linaro.org/lkft/linux-stable-rc-5.6-oe/build/v5.6.7-168-g853ae83af7cc/
  https://qa-reports.linaro.org/lkft/linux-stable-rc-5.6-oe/build/v5.6.7-168-g853ae83af7cc6/
which is very undesirable.

(The other build comes from another source, LKFT 2.0, which is
taking the git-description from Tuxbuild.)

This patch sets the number of hex digits to 12, which seems to
be a prevalent usage and good enough for now. In fact, this
fixes the (unexpected) change of behavior of going to 13-
digits. We can change this to 13 when we so desire by
adjusting this number here.

Change-Id: I4a14cf724a6711b9d34e6ded11a87d206cf0053a
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>