se: tweak CVE reporting emails
Multiple email-ext entries do not seem to work; failed builds are no
longer producing email notifications.

Go back to a single email-ext, that includes the CVE results. For failed
builds this sends CVE results from previous build, which is confusing.

To fix that, run the postbuildscript regardless of build status. In the
script, check the existence of a per-build logfile, to determine if the
build was successful. If logs are missing, create empty CVE report.

Change-Id: I0056d29aa7659d4b288892f7be88e024b596fb49