tcwg/*: Add helper script to generate .yaml files from .yaml.in templates
This allows greater flexibility to parametrize jobs definitions.
The general approach is the same as in ci/dockerfiles.git repo to handle
*-tcwg-* images, and 2 out of 3 scripts are copied verbatim from
ci/dockerfiles.git repo.  Specifically, it is the job of the developer
to run tcwg/generate-yamlfiles.sh script to re-generate all affected .yaml
files and commit them together with the changes to the template.

This patch lays the groundwork to solve [1], and subsequent patch will
enable processing of include files.

[1] https://projects.linaro.org/browse/LSS-576

Change-Id: I2e7b3ab74ae53746b40f60226b7d615b76e76baf