run-jjb.py: Better handle job names with "/" in them
The logic that checks that job name and filename of .yaml file it
came from can be too restrictive for jobs generated using
job-template and .yaml.in templates.  In particular, it is useful
to name template-generated jobs as foo/bar, which both Jenkins and
JJB support.

The motivating case is upcoming rename of tcwg_bmk* jobs to
tcwg_bmk_ci/xxx-<bar>, which will reside in tcwg_bmk_ci_<bar>.yaml files.
Current logic checks that tcwg_bmk_ci doesn't start tcwg_bmk_ci_<bar>
prefix, and reports an error.  This patch reduces the prefix to check
down to "tcwg_bmk_ci", which still provides enough validation that
the job definition is in the right file.

Change-Id: I7c3af6476d121162eeccb56428d4106ac92b5738