tcwg_bmk: Split into 4 configurations and simplify
This change continues to simplify tcwg_bmk* definitions.
We split the single tcwg_bmk.yaml into 4:
- tcwg_bmk_ci_gnu.yaml
- tcwg_bmk_ci_llvm.yaml
- tcwg_bmk_dev_gnu.yaml
- tcwg_bmk_dev_gnu.yaml

This allows us to then split tcwg_bmk.yaml.in to tcwg_bmk_ci.yaml.in,
and tcwg_bmk_dev.yaml.in.

This change also reworks job naming scheme to be more convenient and
scalable.  The old naming scheme was:
- <ci_project>-<action>-<ci_config>
and the new one is:
- <job_template_id>/<action>-<ci_project>/<ci_config>
The new naming scheme will allow further simplifications in TCWG's
round-robin jobs.

Change-Id: I039abba7a1907eb94d8e87e7d03257abec8c3b6f