zephyr-upstream: Remove --link-latest flag to linaro-cp.py.
As we're now publishing to AWS S3, and it doesn't support native
linking operation, it's actually emulated by copying. And as there're
hundreds of files, this "linking" causes noticeable overhead. So,
remove it, as the primary users of these downloads is LAVA anyway,
and on occasional need, humans hopefully can identify the latest
build themselves.

Change-Id: If4227b13090f489d0d8b82617f5008bbd9f4d9fa
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>