ci-docker: dont run in parallel
All these jobs may build/delete the same docker image. We
don't want that to happen in parallel. Else we might end
up with a image built in "sanity check" to be published by
the "deployment" job running on the same node.

This commit assumes the plugin is smart enough not to block itself,
despite the name match.

Change-Id: I8476f467fbc02ca02f17cd75a916984a9cf24611
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>