zephyr-upstream/submit_for_testing: Dump jobdef only in case of error
Previously, we printed each and every jobdef we were about to submit to
LAVA, dozens and hundreds total. This led to very long and verbose build
log dumps in Jenkins. Such information is probably not useful most of
the time, so is now put under the DEBUG_DUMP_JOBDEFS config var, default
False.

However, if there was an error submitting a jobdef, dump its content
together with the error.

Change-Id: I9dda409b0600f2221a60b3e17980068539f8885c
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>