tcwg_bmk,tcwg_gnu,tcwg_kernel: Simplify handling of git URLs in bisect jobs
In the bisect job we need only single git URL of the component that
we are bisecting, which we know when triggering the bisect job.
Therefore, no need to pass git URLs for all other components.

Also, we always bisect vs baseline, so remove the explicit baseline_branch
parameter (we might add it later in developer-friendly versions of
bisect jobs).

Change-Id: Iae0149e1edd3257ded1f403aee59e9df557e1655