lite-aeolus-micropython: Disable not-useful LAVA notifications.
LAVA notifications proved to be not useful. The matter, notifications
are sent unconditionally for each job, regardless of whether it's
successful or not. Then emails are also not very destinguiashable,
requiring effort to figuring out whether it was successful or not,
instead of making it clear right away. All in all, they're indeed
not too useful, and effectively turned into daily spam.

Instead, we're trying to leverage SQUAD's notification capabilities,
and as we're having a kind of progress with untangling setup for
our projects, disable LAVA notifications for both Vincent and myself.

Change-Id: I5f0c198b2fbebb920a08288a0c172bcda5871453
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>