lt-qcom-openembedded-meta-qcom-master-premerge: fix up job
When using git/scm/github integration, Jenkins will automatically checkout
the git tree with the right commit corresponding to the PR. We should
use this checkout insted of fetching again (and as it happens we were
not even fetching the right commit..).

To make things cleaner, move the meta-qcom git in
WORKSPACE/meta-qcom, since we checkout poky in WORKSPACE/poky.

Change-Id: I6fc4f7065c1c2237c3c87f6e0f35c5847c57687b
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>