lt-qcom: remove duplicate failure emails
For these jobs, emails were sent twice in case of failure. As it
happens, 'failure' is set to true by default. So when 'always' is set
to true, we end up with two emails sent:

00:05:10.195 Email was triggered for: Always
00:05:10.195 Email was triggered for: Failure - Any
00:05:10.195 Sending email for trigger: Failure - Any
00:05:11.598 Sending email for trigger: Always

If 'always' is set to true, it looks appropriate to set
'failure' to false explicitely.

Change-Id: Ia8b2f62840bba87c0c62979434e6b7b78201dac5
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>