run-jjb: better support deleting files
Currently the way to delete jobs is setting display-name: "DELETE ME"
in the job you want to delete.

This is not intuitive, but it is required for deleteing TCWG template
jobs.

For everyone else, just support the "D" action, to delete the job when
the file is removed from git.

Change-Id: I06165a7ee8b50097bd35046691aae796f7ef0a97