tcwg-regression-detection: Switch $distro back to xenial
This job triggers tcwg-buildfarm to build gcc-7, which ABE builds with
glibc-2.25.

When we upload the sysroot to the test container, we thus upload
glibc-2.25 but if the test container uses a more recent glibc version,
the remote rsync will fail to start: the rsync shipped with bionic
depends on libpopt.so.0 which in turn requires libc.so.6 with version
GLIBC_2.27.

Change-Id: Ic4f9924e55ce43ad430752b34cb4809a09e7b613