lhg-oe-build/builders.sh: don't fail if *.rootfs.ext4.gz is missing
For some ${MACHINE}'s ext4 rootfs image is not produced. Currently
one needs to add the ${MACHINE} to the list of the ones not to create
the sparse image for - or the build would fail for this only reason.

Let's check that the *.rootfs.ext4.gz does exist before trying to
make the sparse image out of it. This way people still have an option
to disable creating the sparse image for particular ${MACHINE} (regardless
of whether the ext4 image was built or not), and the build would
complete OK if creating the sparse image is not disabled for the
${MACHINE} which doesn't produce the ext4 rootfs image.

Change-Id: Iff3f1f0fd2ced2770d592660bda7948943de2b7b
Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>