leg-kolla: get rid of branch specific stuff
In past we had override per branch as their content was different.
Nowadays they are the same so generic one was created instead for
Train/Ussuri and above.

Branch one can still be created and will be used if present.

Also got rid of branch specific 'devcloud' profiles. One list is now
present as we build Train/Ussuri only.

Change-Id: Ied24859854587e560a168ef3f94fe548eb2855ab