se: disable all yepkit devices before testing
We want all ports off in case they change the test behaviour.

Also, make sure the USB ethernet device is off before we run the NFS
testing. To help with this, I've swapped the WIC and NFS tests so that
WIC comes first.

Throw in a change to use 'run linux_sd' on LCES2.

Change-Id: I88f377231351a2557e236971887b6000a028c045
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>