triggers: move trigger jobs to "triggers" jenkins node
Move triggers off master and build nodes, and use dedicated "triggers"
node instead. The main reason for this is any process running on the
"master" node has all the priviledges of the jenkins process.

Moves from "build" node are not strictly neccessary, but consolidating
them to the same node makes sense as cleanup.

Change-Id: I4f9547c4364ef9857a190f06c463e1d5778ca59f
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>