openemebedded-lkft: Add libgmp-dev to host
The build process started failing on arm 32-bits machines when
the Docker image on which we're building was swapped to a
newer version of Debian.

While adding a dependency to the host in order to fix a build
error is not an idea we're in love with, it is what it is.

Change-Id: I0db23502bd6ceaa61260fa93dff96825252559f5
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>