lkft-generic-build: allow concurrent builds
Currently there is two lkft-generic-build jobs in the que with
the message "pending - a previous build is already in progress"

All lkft-generic-build build happen at the same time to reduce build
waiting time. You may want to apply this to other jobs that are
triggered often at the same time.

Change-Id: I09e4d3765e764d336cf09dd165ca6a03c199f404
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>